+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Added
 - event export with csv and Excel files
+### Updated
+- member overview table is now (more) responsive
 ### Fixed
 - correct working disableLastYearEvents functionality